+++ conftest.py
@@ -1,4 +1,5 @@
 """Root conftest.py"""
 
+# Any tests in developer_resources/ are meant to be run manually.
 # Don't collect e2e tests; only run when specified over CLI.
-collect_ignore = ["tests/e2e_tests"]
+collect_ignore = ["developer_resources", "tests/e2e_tests"]